re-added assert but put NDEBUG to turn them off

+++ src/base_types.hh
@@ -11,10 +11,10 @@ typedef int_least32_t energy_t;
 typedef int_least32_t cand_pos_t;
 typedef uint_fast32_t cand_pos_tu;
 //! type of base pair
-typedef int_least8_t pair_type;
+typedef int_least32_t pair_type;
 //! type of dangle type
 typedef int_least8_t Dangle;
 //! type of base type
-typedef int_least8_t base_type;
+typedef int_least16_t base_type;
 
 #endif // BASE_HH
